chore(dashboard): suggestions from the pr
