+++ apps/dashboard/src/components/subscribers/create-subscriber-form.tsx
@@ -23,7 +23,7 @@ import TruncatedText from '../truncated-text';
 import { LocaleSelect } from './locale-select';
 import { CreateSubscriberFormSchema } from './schema';
 import { TimezoneSelect } from './timezone-select';
-import { useNavigateToSubscribersFirstPage } from './hooks/use-navigate-to-subscribers-first-page';
+import { useSubscribersNavigate } from '@/components/subscribers/hooks/use-subscribers-navigate';
 
 const extensions = [loadLanguage('json')?.extension ?? []];
 const basicSetup = { lineNumbers: true, defaultKeymap: true };
@@ -41,7 +41,7 @@ type CreateSubscriberFormProps = {
 export const CreateSubscriberForm = (props: CreateSubscriberFormProps) => {
   const { onSuccess } = props;
   const track = useTelemetry();
-  const navigateToSubscribersFirstPage = useNavigateToSubscribersFirstPage();
+  const { navigateToSubscribersFirstPage } = useSubscribersNavigate();
   const form = useForm<z.infer<typeof CreateSubscriberFormSchema>>({
     defaultValues: {
       data: '',
@@ -64,6 +64,7 @@ export const CreateSubscriberForm = (props: CreateSubscriberFormProps) => {
       showSuccessToast('Created subscriber successfully', undefined, toastOptions);
       onSuccess?.();
       track(TelemetryEvent.SUBSCRIBER_CREATED);
+      navigateToSubscribersFirstPage();
     },
     onError: (error) => {
       const errMsg = error instanceof Error ? error.message : 'Failed to create subscriber';
@@ -89,7 +90,6 @@ export const CreateSubscriberForm = (props: CreateSubscriberFormProps) => {
     await createSubscriber({
       subscriber: { ...dirtyPayload, subscriberId: formData.subscriberId },
     });
-    navigateToSubscribersFirstPage();
   };
 
   const firstNameChar = form.getValues('firstName')?.charAt(0) || '';
