+++ apps/dashboard/src/components/subscribers/subscriber-overview-form.tsx
@@ -12,7 +12,7 @@ import { loadLanguage } from '@uiw/codemirror-extensions-langs';
 import { useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { RiDeleteBin2Line, RiMailLine } from 'react-icons/ri';
-import { Link, useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { ExternalToast } from 'sonner';
 import { z } from 'zod';
 import { ConfirmationModal } from '../confirmation-modal';
@@ -30,8 +30,7 @@ import { TimezoneSelect } from './timezone-select';
 import { getSubscriberTitle } from './utils';
 import { useSubscribersUrlState } from '@/components/subscribers/hooks/use-subscribers-url-state';
 import { useFetchSubscribers } from '@/hooks/use-fetch-subscribers';
-import { useNavigateToSubscribersFirstPage } from '@/components/subscribers/hooks/use-navigate-to-subscribers-first-page';
-import { useNavigateToSubscribersCurrentPage } from '@/components/subscribers/hooks/use-navigate-to-subscribers-current-page';
+import { useSubscribersNavigate } from '@/components/subscribers/hooks/use-subscribers-navigate';
 
 const extensions = [loadLanguage('json')?.extension ?? []];
 const basicSetup = { lineNumbers: true, defaultKeymap: true };
@@ -63,8 +62,7 @@ export function SubscriberOverviewForm(props: SubscriberOverviewFormProps) {
   const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
   const track = useTelemetry();
 
-  const navigateToSubscribersFirstPage = useNavigateToSubscribersFirstPage();
-  const navigateToSubscribersCurrentPage = useNavigateToSubscribersCurrentPage();
+  const { navigateToSubscribersFirstPage, navigateToSubscribersCurrentPage } = useSubscribersNavigate();
   const { filterValues } = useSubscribersUrlState();
   const { data } = useFetchSubscribers(filterValues, {
     meta: { errorMessage: 'Issue fetching subscribers' },
@@ -74,13 +72,18 @@ export function SubscriberOverviewForm(props: SubscriberOverviewFormProps) {
     onSuccess: () => {
       showSuccessToast(`Deleted subscriber: ${getSubscriberTitle(subscriber)}`, undefined, toastOptions);
       track(TelemetryEvent.SUBSCRIBER_DELETED);
+      const isLastSubscriber = data?.data.length === 1;
+
+      if (isLastSubscriber) {
+        navigateToSubscribersFirstPage();
+      } else {
+        navigateToSubscribersCurrentPage();
+      }
     },
     onError: () => {
       showErrorToast('Failed to delete subscriber', undefined, toastOptions);
     },
   });
-
-  const navigate = useNavigate();
 
   const form = useForm<z.infer<typeof SubscriberFormSchema>>({
     defaultValues: createDefaultSubscriberValues(subscriber),
@@ -409,15 +412,8 @@ export function SubscriberOverviewForm(props: SubscriberOverviewFormProps) {
         open={isDeleteModalOpen}
         onOpenChange={setIsDeleteModalOpen}
         onConfirm={async () => {
-          const isLastSubscriber = data?.data.length === 1;
           await deleteSubscriber({ subscriberId: subscriber.subscriberId });
           setIsDeleteModalOpen(false);
-
-          if (isLastSubscriber) {
-            navigateToSubscribersFirstPage();
-          } else {
-            navigateToSubscribersCurrentPage();
-          }
         }}
         title="Delete subscriber"
         description={
